WIP: refactor v1 protocol
